Switched entire debugging structure to make use of custom functional interfaces. CommandFunction and Draw function are custom functional interfaces created to make this possible. Consider using Guava/Durain’s Errors class to generalize this approach. By returning the debugModule from each function call, the structure is highly chained. To add DebugModules, call something like: Javabot#registerDebugModule("name") 				.addCommand( 						null, 						(c, e) -> { 							e.sendText("Toggled module activity automatically!"); 						}) 				.addAlias("c1") 				.addCommand("c2", 						(c, e) -> e.drawDotMap(150, 150, Color.Black) .setDraw( 					(e) -> { 						e.drawTextScreen(100, 100, 							"Hello world!”); 					}); The above adds a module called name, whose default action is to call the command mapped to null. This command can also be accessed through the alias c1. A second command is mapped to “c2” as well. Finally, this module will draw using the function given in setDraw() if the module is active when the debugger is called from the main loop.
Debugging Submodules are now possible! The command is parsed using a
deterministic FSM, recursively evaluating the current word of the
command. Commands can be input consuming commands, recursively passed to
a submodule, or both. This is all handled by DebugModule’s final
implementation of onReceiveCommand().
The command mapped to “null” in the commands map is called when the
module/submodule is invoked with no arguments.
InvalidCommandException is thrown when a command cannot be parsed.

debug has been moved into its own package under gamestructure.debug.

Many parts of the project are not compile-ready yet.